:racehorse: Use single VAO for multiple objects for perf
Reusing VAOs is useful when models are of the same type.
Meaning that if models are the same type, they will have
the same vertex, indices information. We can utilize
a single VAO that holds this shared information and render
it multiple times, each with a different transformation matrix
to place around the world.